+++ src/auth/oauth/oauth.ts
@@ -14,6 +14,7 @@ import {Session} from '../session';
 import {HttpClient} from '../../clients/http_client/http_client';
 import {DataType} from '../../clients/http_client/types';
 import * as ShopifyErrors from '../../error';
+import {SessionInterface} from '../session/types';
 
 import {
   AuthQuery,
@@ -34,14 +35,14 @@ const ShopifyOAuth = {
    * @param redirect Redirect url for callback
    * @param isOnline Boolean value. If true, appends 'per-user' grant options to authorization url to receive online access token.
    *                 During final oauth request, will receive back the online access token and current online session information.
-   *                 Defaults to offline access.
+   *                 Defaults to online access.
    */
   async beginAuth(
     request: http.IncomingMessage,
     response: http.ServerResponse,
     shop: string,
     redirectPath: string,
-    isOnline = false,
+    isOnline = true,
   ): Promise<string> {
     Context.throwIfUninitialized();
     Context.throwIfPrivateApp('Cannot perform OAuth for private apps');
@@ -55,10 +56,10 @@ const ShopifyOAuth = {
 
     const session = new Session(
       isOnline ? uuidv4() : this.getOfflineSessionId(shop),
+      shop,
+      state,
+      isOnline,
     );
-    session.shop = shop;
-    session.state = state;
-    session.isOnline = isOnline;
 
     const sessionStored = await Context.SESSION_STORAGE.storeSession(session);
 
@@ -99,12 +100,13 @@ const ShopifyOAuth = {
    * @param response Current HTTP Response
    * @param query Current HTTP Request Query, containing the information to be validated.
    *              Depending on framework, this may need to be cast as "unknown" before being passed.
+   * @returns SessionInterface
    */
   async validateAuthCallback(
     request: http.IncomingMessage,
     response: http.ServerResponse,
     query: AuthQuery,
-  ): Promise<void> {
+  ): Promise<SessionInterface> {
     Context.throwIfUninitialized();
     Context.throwIfPrivateApp('Cannot perform OAuth for private apps');
 
@@ -120,7 +122,7 @@ const ShopifyOAuth = {
       );
     }
 
-    const currentSession = await Context.SESSION_STORAGE.loadSession(
+    let currentSession = await Context.SESSION_STORAGE.loadSession(
       sessionCookie,
     );
     if (!currentSession) {
@@ -160,49 +162,47 @@ const ShopifyOAuth = {
       currentSession.expires = sessionExpiration;
       currentSession.scope = scope;
       currentSession.onlineAccessInfo = rest;
+
+      // For an online session in an embedded app, we no longer want the cookie session so we delete it
+      if (Context.IS_EMBEDDED_APP) {
+        // If this is an online session for an embedded app, replace the online session with a JWT session
+        const onlineInfo = currentSession.onlineAccessInfo as OnlineAccessInfo;
+        const jwtSessionId = this.getJwtSessionId(
+          currentSession.shop,
+          `${onlineInfo.associated_user.id}`,
+        );
+        const jwtSession = Session.cloneSession(currentSession, jwtSessionId);
+
+        const sessionDeleted = await Context.SESSION_STORAGE.deleteSession(currentSession.id);
+        if (!sessionDeleted) {
+          throw new ShopifyErrors.SessionStorageError(
+            'OAuth Session could not be deleted. Please check your session storage functionality.',
+          );
+        }
+        currentSession = jwtSession;
+      }
     } else {
+      // Offline sessions (embedded / non-embedded) will use the same id so they don't need to be updated
       const responseBody = postResponse.body as AccessTokenResponse;
       currentSession.accessToken = responseBody.access_token;
       currentSession.scope = responseBody.scope;
     }
 
-    // If this is an offline session, we're no longer interested in the cookie. If it is online in an embedded app, we
-    // want the cookie session to expire a few seconds from now to give the app time to load itself to set up a JWT.
-    // Otherwise, we want to leave the cookie session alone until the actual expiration.
-    let oauthSessionExpiration = currentSession.expires;
-    if (!currentSession.isOnline) {
-      oauthSessionExpiration = new Date();
-    } else if (Context.IS_EMBEDDED_APP) {
-      // If this is an online session for an embedded app, prepare a JWT session to be used going forward
-      const onlineInfo = currentSession.onlineAccessInfo as OnlineAccessInfo;
-      const jwtSessionId = this.getJwtSessionId(
-        currentSession.shop,
-        `${onlineInfo.associated_user.id}`,
-      );
-      const jwtSession = Session.cloneSession(currentSession, jwtSessionId);
-      await Context.SESSION_STORAGE.storeSession(jwtSession);
-
-      // Make sure the current OAuth session expires along with the cookie
-      oauthSessionExpiration = new Date(Date.now() + 30000);
-      currentSession.expires = oauthSessionExpiration;
-    }
-
     cookies.set(ShopifyOAuth.SESSION_COOKIE_NAME, currentSession.id, {
       signed: true,
-      expires: oauthSessionExpiration,
+      expires: Context.IS_EMBEDDED_APP ? new Date() : currentSession.expires,
       sameSite: 'lax',
       secure: true,
     });
 
-    const sessionStored = await Context.SESSION_STORAGE.storeSession(
-      currentSession,
-    );
-
+    const sessionStored = await Context.SESSION_STORAGE.storeSession(currentSession);
     if (!sessionStored) {
       throw new ShopifyErrors.SessionStorageError(
         'OAuth Session could not be saved. Please check your session storage functionality.',
       );
     }
+
+    return currentSession;
   },
 
   /**